requirements.txt
streamlit==1.28.1
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.1.0
langchain-groq>=0.2.0
langchain-huggingface>=0.0.1
faiss-cpu==1.7.4
huggingface-hub>=0.16.0,<1.0
sentence-transformers>=2.7.0
python-dotenv==1.0.0
pypdf>=4.0.1
